Conversion to double Explicitly template to SparseVector Remove id map requirements Various clean up (removing commented out & dead code)

+++ src/main/java/edu/berkeley/compbio/jlibsvm/ContinuousModel.java
@@ -11,5 +11,5 @@ package edu.berkeley.compbio.jlibsvm;
 public interface ContinuousModel<P> {
 // -------------------------- OTHER METHODS --------------------------
 
-  Float predictValue(P x);
+  Double predictValue(P x);
 }
